Updated UserRating in Rating.cs to be of decimal type and increment by 0.5 from 0 to 5

+++ TeamWebApplication/Repositories/Interfaces/IRatingsRepository.cs
@@ -10,6 +10,6 @@ namespace TeamWebApplication.Repositories.Interfaces
         Task DeleteRatingAsync(Rating? rating);
         Task UpdateRatingAsync(Rating? rating);
         Task UpdateRatingAsync(Rating? originalRating, Rating? rating);
-        Task<double> GetCourseRatingAsync(int? courseId);
+        Task<decimal> GetCourseRatingAsync(int? courseId);
     }
 }
